refactor: no separate boot procedure
