+++ src/core/components/add/index.js
@@ -6,11 +6,12 @@ const { parseChunkerString } = require('./utils')
 const pipe = require('it-pipe')
 
 module.exports = ({ ipld, dag, gcLock, preload, pin, constructorOptions }) => {
+  const isShardingEnabled = constructorOptions.EXPERIMENTAL && constructorOptions.EXPERIMENTAL.sharding
   return async function * add (source, options) {
     options = options || {}
 
     const opts = {
-      shardSplitThreshold: constructorOptions.EXPERIMENTAL.sharding ? 1000 : Infinity,
+      shardSplitThreshold: isShardingEnabled ? 1000 : Infinity,
       ...options,
       strategy: 'balanced',
       ...parseChunkerString(options.chunker)
