+++ src/core/components/index.js
@@ -1,31 +1,14 @@
 'use strict'
 
-exports.preStart = require('./pre-start')
+exports.add = require('./add')
+exports.init = require('./init')
 exports.start = require('./start')
 exports.stop = require('./stop')
-exports.isOnline = require('./is-online')
-exports.version = require('./version')
-exports.id = require('./id')
-exports.repo = require('./repo')
-exports.init = require('./init')
-exports.bootstrap = require('./bootstrap')
-exports.config = require('./config')
-exports.block = require('./block')
-exports.object = require('./object')
-exports.dag = require('./dag')
-exports.libp2p = require('./libp2p')
-exports.swarm = require('./swarm')
-exports.ping = require('./ping')
-exports.pingPullStream = require('./ping-pull-stream')
-exports.pingReadableStream = require('./ping-readable-stream')
-exports.pin = require('./pin')
-exports.filesRegular = require('./files-regular')
-exports.filesMFS = require('./files-mfs')
-exports.bitswap = require('./bitswap')
-exports.pubsub = require('./pubsub')
-exports.dht = require('./dht')
-exports.dns = require('./dns')
-exports.key = require('./key')
-exports.stats = require('./stats')
-exports.resolve = require('./resolve')
-exports.name = require('./name')
+
+exports.legacy = { // TODO: these will be removed as the new API is completed
+  config: require('./config'),
+  dag: require('./dag'),
+  libp2p: require('./libp2p'),
+  object: require('./object'),
+  pin: require('./pin')
+}
