+++ src/core/components/init.js
@@ -1,164 +1,307 @@
 'use strict'
 
-const peerId = require('peer-id')
+const log = require('debug')('ipfs:components:init')
+const PeerId = require('peer-id')
+const PeerInfo = require('peer-info')
 const mergeOptions = require('merge-options')
-const callbackify = require('callbackify')
 const promisify = require('promisify-es6')
-const defaultConfig = require('../runtime/config-nodejs.js')
+const getDefaultConfig = require('../runtime/config-nodejs.js')
+const createRepo = require('../runtime/repo-nodejs')
 const Keychain = require('libp2p-keychain')
-const {
-  DAGNode
-} = require('ipld-dag-pb')
+const NoKeychain = require('./no-keychain')
+const GCLock = require('./pin/gc-lock')
+const { DAGNode } = require('ipld-dag-pb')
 const UnixFs = require('ipfs-unixfs')
 const multicodec = require('multicodec')
-
+const multiaddr = require('multiaddr')
+const {
+  AlreadyInitializingError,
+  AlreadyInitializedError,
+  NotStartedError
+} = require('../errors')
+const BlockService = require('ipfs-block-service')
+const Ipld = require('ipld')
+const getDefaultIpldOptions = require('../runtime/ipld-nodejs')
+const createPreloader = require('../preload')
+const { ERR_REPO_NOT_INITIALIZED } = require('ipfs-repo').errors
 const IPNS = require('../ipns')
 const OfflineDatastore = require('../ipns/routing/offline-datastore')
+const initAssets = require('../runtime/init-assets-nodejs')
+const PinManager = require('./pin/pin-manager')
+const Commands = require('./')
 
-const addDefaultAssets = require('./init-assets')
-const { profiles } = require('./config')
+module.exports = ({
+  apiManager,
+  print,
+  constructorOptions
+}) => async function init (options) {
+  const { cancel } = apiManager.update({ init: () => { throw new AlreadyInitializingError() } })
 
-function createPeerId (self, opts) {
-  if (opts.privateKey) {
-    self.log('using user-supplied private-key')
-    if (typeof opts.privateKey === 'object') {
-      return opts.privateKey
-    } else {
-      return promisify(peerId.createFromPrivKey)(Buffer.from(opts.privateKey, 'base64'))
+  try {
+    options = options || {}
+
+    if (typeof constructorOptions.init === 'object') {
+      options = mergeOptions(options, constructorOptions.init)
     }
-  } else {
-    // Generate peer identity keypair + transform to desired format + add to config.
-    opts.log(`generating ${opts.bits}-bit RSA keypair...`, false)
-    self.log('generating peer id: %s bits', opts.bits)
 
-    return promisify(peerId.create)({ bits: opts.bits })
+    if (constructorOptions.pass) {
+      options.pass = constructorOptions.pass
+    }
+
+    if (constructorOptions.config) {
+      options.config = constructorOptions.config
+    }
+
+    const repo = typeof options.repo === 'string' || options.repo == null
+      ? createRepo({ path: options.repo, autoMigrate: options.repoAutoMigrate })
+      : options.repo
+
+    let isInitialized = true
+
+    if (repo.closed) {
+      try {
+        await repo.open()
+      } catch (err) {
+        if (err.code === ERR_REPO_NOT_INITIALIZED) {
+          isInitialized = false
+        } else {
+          throw err
+        }
+      }
+    }
+
+    const { peerId, config, keychain } = isInitialized
+      ? await initExistingRepo(repo, options)
+      : await initNewRepo(repo, options)
+
+    log('peer created')
+    const peerInfo = new PeerInfo(peerId)
+
+    if (config.Addresses && config.Addresses.Swarm) {
+      config.Addresses.Swarm.forEach(addr => {
+        let ma = multiaddr(addr)
+
+        if (ma.getPeerId()) {
+          ma = ma.encapsulate(`/p2p/${peerInfo.id.toB58String()}`)
+        }
+
+        peerInfo.multiaddrs.add(ma)
+      })
+    }
+
+    const blockService = new BlockService(repo)
+    const ipld = new Ipld(getDefaultIpldOptions(blockService, constructorOptions.ipld, log))
+
+    const preload = createPreloader(constructorOptions.preload)
+    await preload.start()
+
+    const gcLock = new GCLock(constructorOptions.repoOwner, {
+      // Make sure GCLock is specific to repo, for tests where there are
+      // multiple instances of IPFS
+      morticeId: repo.path
+    })
+
+    const dag = Commands.legacy.dag({ _ipld: ipld, _preload: preload })
+    const object = Commands.legacy.object({ _ipld: ipld, _preload: preload, dag, _gcLock: gcLock })
+
+    const pinManager = new PinManager(repo, dag)
+    await pinManager.load()
+
+    const pin = Commands.legacy.pin({ _ipld: ipld, _preload: preload, object, _repo: repo, _pinManager: pinManager })
+    const add = Commands.add({ ipld, dag, preload, pin, gcLock, constructorOptions })
+
+    if (!isInitialized && !options.emptyRepo) {
+      // add empty unixfs dir object (go-ipfs assumes this exists)
+      const emptyDirCid = await addEmptyDir({ dag })
+
+      log('adding default assets')
+      await initAssets({ add, print })
+
+      log('initializing IPNS keyspace')
+      // Setup the offline routing for IPNS.
+      // This is primarily used for offline ipns modifications, such as the initializeKeyspace feature.
+      const offlineDatastore = new OfflineDatastore(repo)
+      const ipns = new IPNS(offlineDatastore, repo.datastore, peerInfo, keychain, { pass: options.pass })
+      await ipns.initializeKeyspace(peerId.privKey.bytes, emptyDirCid.toString())
+    }
+
+    const api = createApi({
+      add,
+      apiManager,
+      constructorOptions,
+      blockService,
+      gcLock,
+      initOptions: options,
+      ipld,
+      keychain,
+      peerInfo,
+      pinManager,
+      preload,
+      print,
+      repo
+    })
+
+    apiManager.update(api, () => { throw new NotStartedError() })
+  } catch (err) {
+    cancel()
+    throw err
   }
+
+  return apiManager.api
 }
 
-async function createRepo (self, opts) {
-  if (self.state.state() !== 'uninitialized') {
-    throw new Error('Not able to init from state: ' + self.state.state())
-  }
+async function initNewRepo (repo, { privateKey, emptyRepo, bits, profiles, config, pass, print }) {
+  emptyRepo = emptyRepo || false
+  bits = bits == null ? 2048 : Number(bits)
 
-  self.state.init()
-  self.log('init')
-
-  // An initialized, open repo was passed, use this one!
-  if (opts.repo) {
-    self._repo = opts.repo
-
-    return
-  }
-
-  opts.emptyRepo = opts.emptyRepo || false
-  opts.bits = Number(opts.bits) || 2048
-  opts.log = opts.log || function () {}
-
-  const config = mergeOptions(defaultConfig(), self._options.config)
-
-  applyProfile(self, config, opts)
+  config = mergeOptions(getDefaultConfig(), config)
+  config = applyProfiles(profiles, config)
 
   // Verify repo does not exist yet
-  const exists = await self._repo.exists()
-  self.log('repo exists?', exists)
+  const exists = await repo.exists()
+  log('repo exists?', exists)
+
   if (exists === true) {
-    throw Error('repo already exists')
+    throw new Error('repo already exists')
   }
 
-  const peerId = await createPeerId(self, opts)
+  const peerId = await createPeerId({ privateKey, bits, print })
+  let keychain = new NoKeychain()
 
-  self.log('identity generated')
+  log('identity generated')
 
   config.Identity = {
     PeerID: peerId.toB58String(),
     PrivKey: peerId.privKey.bytes.toString('base64')
   }
-  const privateKey = peerId.privKey
 
-  if (opts.pass) {
-    config.Keychain = Keychain.generateOptions()
+  privateKey = peerId.privKey
+
+  config.Keychain = Keychain.generateOptions()
+
+  log('peer identity: %s', config.Identity.PeerID)
+
+  await repo.init(config)
+  await repo.open()
+
+  log('repo opened')
+
+  if (pass) {
+    log('creating keychain')
+    const keychainOptions = { passPhrase: pass, ...config.Keychain }
+    keychain = new Keychain(repo.keys, keychainOptions)
+    await keychain.importPeer('self', { privKey: privateKey })
   }
 
-  opts.log('done')
-  opts.log('peer identity: ' + config.Identity.PeerID)
-
-  await self._repo.init(config)
-  await self._repo.open()
-
-  self.log('repo opened')
-
-  if (opts.pass) {
-    self.log('creating keychain')
-    const keychainOptions = Object.assign({ passPhrase: opts.pass }, config.Keychain)
-    self._keychain = new Keychain(self._repo.keys, keychainOptions)
-
-    await self._keychain.importPeer('self', { privKey: privateKey })
-  }
-
-  // Setup the offline routing for IPNS.
-  // This is primarily used for offline ipns modifications, such as the initializeKeyspace feature.
-  const offlineDatastore = new OfflineDatastore(self._repo)
-
-  self._ipns = new IPNS(offlineDatastore, self._repo.datastore, self._peerInfo, self._keychain, self._options)
-
-  // add empty unixfs dir object (go-ipfs assumes this exists)
-  return addRepoAssets(self, privateKey, opts)
+  return { peerId, keychain, config }
 }
 
-async function addRepoAssets (self, privateKey, opts) {
-  if (opts.emptyRepo) {
-    return
+async function initExistingRepo (repo, { config: newConfig, profiles, pass }) {
+  let config = await repo.config.get()
+
+  if (newConfig || profiles) {
+    if (newConfig) {
+      config = mergeOptions(config, newConfig)
+    }
+    if (profiles) {
+      config = applyProfiles(profiles, config)
+    }
+    await repo.config.set(config)
   }
 
-  self.log('adding assets')
+  let keychain = new NoKeychain()
 
+  if (pass) {
+    const keychainOptions = { passPhrase: pass, ...config.Keychain }
+    keychain = new Keychain(repo.keys, keychainOptions)
+    log('keychain constructed')
+  }
+
+  const peerId = await promisify(PeerId.createFromPrivKey)(config.Identity.PrivKey)
+
+  // Import the private key as 'self', if needed.
+  if (pass) {
+    try {
+      await keychain.findKeyByName('self')
+    } catch (err) {
+      log('Creating "self" key')
+      await keychain.importPeer('self', peerId)
+    }
+  }
+
+  return { peerId, keychain, config }
+}
+
+function createPeerId ({ privateKey, bits, print }) {
+  if (privateKey) {
+    log('using user-supplied private-key')
+    return typeof privateKey === 'object'
+      ? privateKey
+      : promisify(PeerId.createFromPrivKey)(Buffer.from(privateKey, 'base64'))
+  } else {
+    // Generate peer identity keypair + transform to desired format + add to config.
+    print('generating %s-bit RSA keypair...', bits)
+    return promisify(PeerId.create)({ bits })
+  }
+}
+
+async function addEmptyDir ({ dag }) {
   const node = new DAGNode(new UnixFs('directory').marshal())
-  const cid = await self.dag.put(node, {
+  return dag.put(node, {
     version: 0,
     format: multicodec.DAG_PB,
     hashAlg: multicodec.SHA2_256,
     preload: false
   })
-
-  await self._ipns.initializeKeyspace(privateKey, cid.toBaseEncodedString())
-
-  self.log('Initialised keyspace')
-
-  if (typeof addDefaultAssets === 'function') {
-    self.log('Adding default assets')
-    // addDefaultAssets is undefined on browsers.
-    // See package.json browser config
-    return addDefaultAssets(self, opts.log)
-  }
 }
 
-// Apply profiles (eg "server,lowpower") to config
-function applyProfile (self, config, opts) {
-  if (opts.profiles) {
-    for (const name of opts.profiles) {
-      const profile = profiles[name]
-
-      if (!profile) {
-        throw new Error(`Could not find profile with name '${name}'`)
-      }
-
-      self.log(`applying profile ${name}`)
-      profile.transform(config)
+// Apply profiles (e.g. ['server', 'lowpower']) to config
+function applyProfiles (profiles, config) {
+  return (profiles || []).reduce((name, config) => {
+    const profile = require('./config').profiles[name]
+    if (!profile) {
+      throw new Error(`No profile with name '${name}'`)
     }
-  }
+    log('applying profile %s', name)
+    return profile.transform(config)
+  })
 }
 
-module.exports = function init (self) {
-  return callbackify.variadic(async (opts) => {
-    opts = opts || {}
-
-    await createRepo(self, opts)
-    self.log('Created repo')
-
-    await self.preStart()
-    self.log('Done pre-start')
-
-    self.state.initialized()
-    self.emit('init')
+function createApi ({
+  add,
+  apiManager,
+  constructorOptions,
+  blockService,
+  gcLock,
+  initOptions,
+  ipld,
+  keychain,
+  peerInfo,
+  pinManager,
+  preload,
+  print,
+  repo
+}) {
+  const start = Commands.start({
+    apiManager,
+    constructorOptions,
+    blockService,
+    gcLock,
+    initOptions,
+    ipld,
+    keychain,
+    peerInfo,
+    pinManager,
+    preload,
+    print,
+    repo
   })
+
+  const api = {
+    add,
+    init: () => { throw new AlreadyInitializedError() },
+    start
+  }
+
+  return api
 }
