+++ src/core/runtime/repo-browser.js
@@ -3,6 +3,7 @@
 const IPFSRepo = require('ipfs-repo')
 
 module.exports = (options) => {
-  const repoPath = options.repo || 'ipfs'
-  return new IPFSRepo(repoPath, { autoMigrate: options.repoAutoMigrate })
+  options = options || {}
+  const repoPath = options.path || 'ipfs'
+  return new IPFSRepo(repoPath, { autoMigrate: options.autoMigrate })
 }
