+++ src/core/runtime/repo-nodejs.js
@@ -4,8 +4,8 @@ const os = require('os')
 const IPFSRepo = require('ipfs-repo')
 const path = require('path')
 
-module.exports = (options) => {
-  const repoPath = options.repo || path.join(os.homedir(), '.jsipfs')
-
-  return new IPFSRepo(repoPath, { autoMigrate: options.repoAutoMigrate })
+module.exports = options => {
+  options = options || {}
+  const repoPath = options.path || path.join(os.homedir(), '.jsipfs')
+  return new IPFSRepo(repoPath, { autoMigrate: options.autoMigrate })
 }
